requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
chromadb==0.4.22
sentence-transformers==2.2.2

# PDF processing
pdfplumber==0.10.0
pymupdf==1.23.14
python-dotenv==1.0.0

# Document processing and NLP
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2

# Vector operations and similarity
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4

# Web interface
flask==2.3.3
flask-cors==4.0.0
plotly==5.17.0

# Google Gemini API
google-generativeai==0.3.2

# Utilities
pandas==2.1.4
requests==2.31.0
tqdm==4.66.1